Began creating cogs.
Turns out cogs are a think, need to acutally make it load cogs.
